requirements.txt
beautifulsoup4==4.3.2
certifi==2015.11.20.1
cffi==1.3.1
cryptography==1.1.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.7
django-durationfield==0.5.1
django-tastypie==0.12.2
django-toolbelt==0.0.1
enum34==1.1.1
gunicorn==19.4.1
idna==2.0
ipaddress==1.0.15
ndg-httpsclient==0.4.0
psycopg2==2.6.1
pyasn1==0.1.9
pycparser==2.14
pyOpenSSL==0.15.1
python-dateutil==2.4.2
python-mimeparse==0.1.4
requests==2.8.1
six==1.10.0
static3==0.6.1
tldextract==1.6
unicodecsv==0.13.0
wheel==0.24.0
